# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Frameless Creative Admin API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /filmmaking-documents/{id}/submit:
    post:
      operationId: submitDocumentToAdmin
      tags: [filmmaking-submissions]
      summary: Submit document to admin for review
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Document submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilmmakingSubmission"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /filmmaking-submissions:
    get:
      operationId: listFilmmakingSubmissions
      tags: [filmmaking-submissions]
      summary: List all submissions (admin only)
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, approved, revision_requested]
          description: Filter by submission status
      responses:
        "200":
          description: List of submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FilmmakingSubmissionDetail"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /filmmaking-submissions/{id}:
    get:
      operationId: getFilmmakingSubmission
      tags: [filmmaking-submissions]
      summary: Get submission details
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Submission details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilmmakingSubmissionFull"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /filmmaking-submissions/{id}/approve:
    post:
      operationId: approveFilmmakingSubmission
      tags: [filmmaking-submissions]
      summary: Approve submission (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Submission approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilmmakingSubmission"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /filmmaking-submissions/{id}/request-revision:
    post:
      operationId: requestFilmmakingRevision
      tags: [filmmaking-submissions]
      summary: Request revision (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestRevisionBody"
      responses:
        "200":
          description: Revision requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilmmakingSubmission"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /filmmaking-documents/{documentId}/export-pdf:
    post:
      operationId: exportFilmmakingDocumentPdf
      tags: [filmmaking-export]
      summary: Export document as PDF (printable HTML)
      parameters:
        - in: path
          name: documentId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: HTML content ready for printing to PDF
          content:
            text/html:
              schema:
                type: string
        "404":
          description: Document not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
...
    FilmmakingSubmission:
      type: object
      properties:
        id:
          type: string
        documentId:
          type: string
        submittedBy:
          type: string
        submittedAt:
          type: string
        status:
          type: string
          enum: [pending, approved, revision_requested]
        adminNotes:
          type: string
          nullable: true
        approvedAt:
          type: string
          nullable: true
        approvedBy:
          type: string
          nullable: true
      required: [id, documentId, submittedBy, submittedAt, status]
    FilmmakingSubmissionDetail:
      allOf:
        - $ref: "#/components/schemas/FilmmakingSubmission"
        - type: object
          properties:
            documentTitle:
              type: string
              nullable: true
            documentType:
              type: string
              nullable: true
            crewMemberName:
              type: string
              nullable: true
            crewMemberEmail:
              type: string
              nullable: true
    FilmmakingSubmissionFull:
      allOf:
        - $ref: "#/components/schemas/FilmmakingSubmission"
        - type: object
          properties:
            document:
              type: object
              properties:
                id:
                  type: string
                title:
                  type: string
                docType:
                  type: string
                content:
                  type: object
                projectId:
                  type: string
                  nullable: true
                createdAt:
                  type: string
                updatedAt:
                  type: string
            crewMember:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
                email:
                  type: string
                  nullable: true
                avatarUrl:
                  type: string
                  nullable: true
    RequestRevisionBody:
      type: object
      properties:
        adminNotes:
          type: string
          nullable: true
...